cloud/upload/model/requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.3
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
image-slicer==0.3.0
imageio==2.6.1
imutils==0.5.3
isort==4.3.21
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
matplotlib==3.1.1
mccabe==0.6.1
networkx==2.4
numpy==1.17.3
oauthlib==3.1.0
opencv-contrib-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.3
patchify==0.1.1
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pylint==2.4.3
pyparsing==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
svgwrite==1.3.1
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.38.0
Tree==0.2.4
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2